requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
cycler==0.11.0
Cython==3.0.1
gast==0.4.0
grpcio==1.57.0
h5py==3.9.0
importlib-metadata==6.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.5
matplotlib==3.7.2
mock==5.1.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
protobuf==4.24.3
pyparsing==3.0.9
python-dateutil==2.8.2
pygame==2.5.0
six==1.16.0
tensorboard==2.13.0
tensorflow-macos==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.3.0
typing-extensions==4.5.0  # Or upgrade to >=4.6.0 if not using tensorflow-macos
sqlalchemy==1.4.46  # Downgrade for compatibility with typing-extensions 4.5.0
Werkzeug==3.0.1
zipp==3.16.2